Moved all tests to a single folder in rootdir
