+++ tests/unit/smartcontracts/pointSocial.ts
